jdk11 fix SSL exceptions
git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@3972 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/ui/musiccollection/MusicCollection.java
@@ -781,7 +781,7 @@ public class MusicCollection extends C64VBox implements UIPart {
 							downloadedFile.delete();
 						}
 						downloadStart(url, hvscVersion + 1);
-					} else {
+					} else if (downloadedFile != null) {
 						downloadedFile.deleteOnExit();
 						Platform.runLater(() -> {
 							Configuration config = util.getConfig();
